requirements.txt
python-dotenv
pandas
numpy
matplotlib
scikit-learn
pymongo
seaborn
flask
dill
certifi
pymongo
pyaml
mlflow
# -e .